Updated ReEnroll
Added Device ReEnroll Method Extension Attribute. Added a check to see if the Dialog window was closed prematurely to give a notice the process is complete. Added infotext in the dialog window to show script version and if in debug mode. Added more the property list creation. Added a root check of the script. Added a check for if the computer if a user is logged in, checking for the dock and finder.